fix: docs & api edge cases
