+++ crates/rattler_repodata_gateway/src/fetch/jlap/mod.rs
@@ -102,7 +102,8 @@ use tempfile::NamedTempFile;
 use tokio::task::JoinError;
 use url::Url;
 
-pub use crate::fetch::cache::{CacheState, JLAPFooter, JLAPState};
+pub use crate::fetch::cache::{JLAPFooter, JLAPState};
+pub use crate::fetch::RepoDataState;
 
 /// File suffix for JLAP file
 pub const JLAP_FILE_SUFFIX: &str = "jlap";
@@ -412,7 +413,7 @@ fn get_bytes_offset(lines: &Vec<&str>) -> u64 {
 pub async fn patch_repo_data(
     client: &ClientWithMiddleware,
     subdir_url: Url,
-    repo_data_state: CacheState,
+    repo_data_state: RepoDataState,
     repo_data_json_path: &Path,
 ) -> Result<(JLAPState, Blake2b256Hash), JLAPError> {
     // Determine what we should use as our starting state
@@ -634,7 +635,7 @@ mod test {
     use super::patch_repo_data;
     use std::path::PathBuf;
 
-    use crate::fetch::cache::CacheState;
+    use crate::fetch::RepoDataState;
     use crate::utils::simple_channel_server::SimpleChannelServer;
 
     use rattler_digest::{parse_digest_from_hex, Blake2b256};
@@ -820,7 +821,7 @@ mod test {
 
         pub client: ClientWithMiddleware,
 
-        pub repo_data_state: CacheState,
+        pub repo_data_state: RepoDataState,
     }
 
     impl TestEnvironment {
@@ -834,7 +835,7 @@ mod test {
 
             let client = reqwest::Client::new().into();
 
-            let repo_data_state: CacheState = serde_json::from_str(repo_data_state).unwrap();
+            let repo_data_state: RepoDataState = serde_json::from_str(repo_data_state).unwrap();
 
             Self {
                 _server: server,
